LayoutTests/fast/dom/Range/range-remove-children-event-expected.txt

Test ranges remain valid when modified inside a MutationEvent handler.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS range.startContainer.tagName is "DIV"
PASS successfullyParsed is true

TEST COMPLETE
